+++ pkg/controller/sync/dispatch/managed.go
@@ -17,6 +17,7 @@ limitations under the License.
 package dispatch
 
 import (
+	"context"
 	"fmt"
 	"strings"
 	"sync"
@@ -24,11 +25,12 @@ import (
 	"github.com/pkg/errors"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	pkgruntime "k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/util/sets"
 
+	"sigs.k8s.io/kubefed/pkg/client/generic"
 	"sigs.k8s.io/kubefed/pkg/controller/sync/status"
 	"sigs.k8s.io/kubefed/pkg/controller/util"
 )
@@ -38,6 +40,7 @@ import (
 type FederatedResourceForDispatch interface {
 	TargetName() util.QualifiedName
 	TargetKind() string
+	TargetGVK() schema.GroupVersionKind
 	Object() *unstructured.Unstructured
 	VersionForCluster(clusterName string) (string, error)
 	ObjectForCluster(clusterName string) (*unstructured.Unstructured, error)
@@ -80,7 +83,7 @@ func NewManagedDispatcher(clientAccessor clientAccessorFunc, fedResource Federat
 		skipAdoptingResources: skipAdoptingResources,
 	}
 	d.dispatcher = newOperationDispatcher(clientAccessor, d)
-	d.unmanagedDispatcher = newUnmanagedDispatcher(d.dispatcher, d, fedResource.TargetKind(), fedResource.TargetName())
+	d.unmanagedDispatcher = newUnmanagedDispatcher(d.dispatcher, d, fedResource.TargetGVK(), fedResource.TargetName())
 	return d
 }
 
@@ -126,7 +129,7 @@ func (d *managedDispatcherImpl) Create(clusterName string) {
 
 	d.dispatcher.incrementOperationsInitiated()
 	const op = "create"
-	go d.dispatcher.clusterOperation(clusterName, op, func(client util.ResourceClient) util.ReconciliationStatus {
+	go d.dispatcher.clusterOperation(clusterName, op, func(client generic.Client) util.ReconciliationStatus {
 		d.recordEvent(clusterName, op, "Creating")
 
 		obj, err := d.fedResource.ObjectForCluster(clusterName)
@@ -139,9 +142,9 @@ func (d *managedDispatcherImpl) Create(clusterName string) {
 			return d.recordOperationError(status.ApplyOverridesFailed, clusterName, op, err)
 		}
 
-		createdObj, err := client.Resources(obj.GetNamespace()).Create(obj, metav1.CreateOptions{})
+		err = client.Create(context.Background(), obj)
 		if err == nil {
-			version := util.ObjectVersion(createdObj)
+			version := util.ObjectVersion(obj)
 			d.recordVersion(clusterName, version)
 			return util.StatusAllOK
 		}
@@ -155,19 +158,19 @@ func (d *managedDispatcherImpl) Create(clusterName string) {
 
 		// Attempt to update the existing resource to ensure that it
 		// is labeled as a managed resource.
-		clusterObj, err := client.Resources(obj.GetNamespace()).Get(obj.GetName(), metav1.GetOptions{})
+		err = client.Get(context.Background(), obj, obj.GetNamespace(), obj.GetName())
 		if err != nil {
 			wrappedErr := errors.Wrapf(err, "failed to retrieve object potentially requiring adoption")
 			return d.recordOperationError(status.RetrievalFailed, clusterName, op, wrappedErr)
 		}
 
-		if d.skipAdoptingResources && !d.fedResource.IsNamespaceInHostCluster(clusterObj) {
+		if d.skipAdoptingResources && !d.fedResource.IsNamespaceInHostCluster(obj) {
 			_ = d.recordOperationError(status.AlreadyExists, clusterName, op, errors.Errorf("Resource pre-exist in cluster"))
 			return util.StatusAllOK
 		}
 
 		d.recordError(clusterName, op, errors.Errorf("An update will be attempted instead of a creation due to an existing resource"))
-		d.Update(clusterName, clusterObj)
+		d.Update(clusterName, obj)
 		return util.StatusAllOK
 	})
 }
@@ -177,7 +180,7 @@ func (d *managedDispatcherImpl) Update(clusterName string, clusterObj *unstructu
 
 	d.dispatcher.incrementOperationsInitiated()
 	const op = "update"
-	go d.dispatcher.clusterOperation(clusterName, op, func(client util.ResourceClient) util.ReconciliationStatus {
+	go d.dispatcher.clusterOperation(clusterName, op, func(client generic.Client) util.ReconciliationStatus {
 		obj, err := d.fedResource.ObjectForCluster(clusterName)
 		if err != nil {
 			return d.recordOperationError(status.ComputeResourceFailed, clusterName, op, err)
@@ -206,11 +209,11 @@ func (d *managedDispatcherImpl) Update(clusterName string, clusterObj *unstructu
 		// Only record an event if the resource is not current
 		d.recordEvent(clusterName, op, "Updating")
 
-		updatedObj, err := client.Resources(obj.GetNamespace()).Update(obj, metav1.UpdateOptions{})
+		err = client.Update(context.Background(), obj)
 		if err != nil {
 			return d.recordOperationError(status.UpdateFailed, clusterName, op, err)
 		}
-		version = util.ObjectVersion(updatedObj)
+		version = util.ObjectVersion(obj)
 		d.recordVersion(clusterName, version)
 		return util.StatusAllOK
 	})